+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Version 0.2 (development)
 
+- BREAKING CHANGE: Rename api.Server class to api.Client
 - Add support to changing client ids (switch between Web and Android)
 - Fix login demo
-- BREAKING CHANGE: Rename api.Server class to api.Client
+- Add Attendance data support